Create Rio Escondido
https://arcg.is/0umD0e0

I have not taken a vacation in a long time with 3 kids 3 and under we do not get many opportunities, and I did not have time before that with work so that task would not fit well with my lifestyle. I tried to replicate that through where I live. I hope this is adequate.